// webapp/src/client/api/generated/model/personalTransactionRead.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Marumie Backend
 * OpenAPI spec version: 0.1.0
 */
import type { PersonalTransactionReadSubcategory } from "./personalTransactionReadSubcategory";
import type { PersonalTransactionType } from "./personalTransactionType";
import type { PersonalTransactionReadMemo } from "./personalTransactionReadMemo";
import type { PersonalTransactionReadOrganizationId } from "./personalTransactionReadOrganizationId";

export interface PersonalTransactionRead {
  id: string;
  date: string;
  category: string;
  subcategory?: PersonalTransactionReadSubcategory;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  amount: string;
  type: PersonalTransactionType;
  payment_method: string;
  description: string;
  memo?: PersonalTransactionReadMemo;
  hash: string;
  created_at: string;
  updated_at: string;
  organization_id?: PersonalTransactionReadOrganizationId;
}